Moved provider accommodation id model back
