+++ Api/Models/Accommodations/ReportAccommodationDuplicateRequest.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using HappyTravel.Edo.Data.AccommodationMappings;
 using Newtonsoft.Json;
 
 namespace HappyTravel.Edo.Api.Models.Accommodations
