+++ Api/Services/Accommodations/Availability/AvailabilityStorage.cs
@@ -11,7 +11,6 @@ using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Api.Models.Availabilities;
 using HappyTravel.Edo.Api.Services.Accommodations.Mappings;
 using HappyTravel.Edo.Common.Enums;
-using HappyTravel.Edo.Data.AccommodationMappings;
 using HappyTravel.EdoContracts.Accommodations;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
